FINISH: eth_call ut && error check
